fix: inline EventHelper concern into Event model
EventHelper was a single-model concern (only used by Event). Inline its
validation methods directly into Event and remove the concern file to
eliminate unnecessary indirection.

This follows the precedent set by inlining Permissions into Member.
